requirements.txt
aio-pika==9.0.4
aiormq==6.7.3
alembic==1.10.1
anyio==3.6.2
asyncio==3.4.3
cfgv==3.3.1
click==8.1.3
distlib==0.3.6
fastapi==0.92.0
filelock==3.9.0
greenlet==2.0.2
h11==0.14.0
identify==2.5.18
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
nodeenv==1.7.0
pamqp==3.2.1
pika==1.3.1
platformdirs==3.1.0
pre-commit==3.1.1
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.5.post1
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
virtualenv==20.20.0
yarl==1.8.2
